--- website/pages/blog/_meta.ts
@@ -1,13 +1,0 @@
-export default {
-  '*': {
-    display: 'hidden',
-    theme: {
-      layout: 'default',
-      sidebar: false,
-      toc: true,
-      typesetting: 'article',
-      breadcrumb: false,
-      pagination: false,
-    },
-  },
-};
