+++ website/theme.config.tsx
@@ -1,15 +1,11 @@
-import { ReactElement } from 'react';
-import { useRouter } from 'next/router';
-import { Article } from '@/article';
-import { BlogCardList, Newsletter, Video } from '@/components';
+import { Video } from '@/components';
 import { CodeSandbox } from '@/shared/embed/code-sandbox';
 import { Gfycat } from '@/shared/embed/gfycat';
 import { LinkPreview } from '@/shared/embed/link-preview';
 import { OgCard } from '@/shared/embed/og-card';
 import { StackBlitz } from '@/shared/embed/stack-blitz';
 import { Tweet } from '@/shared/embed/tweet';
-import { Callout, defineConfig, Giscus, Steps, useConfig, useTheme } from '@theguild/components';
-import { allBlogs } from './lib/all-blogs';
+import { Callout, defineConfig, Steps, useConfig } from '@theguild/components';
 
 function ensureAbsolute(url: string): string {
   return url.startsWith('/') ? `https://the-guild.dev${url}` : url;
@@ -50,55 +46,6 @@ export default defineConfig({
           )}
         />
         <meta property="og:site_name" content={siteName} />
-      </>
-    );
-  },
-  main: function Main({ children }) {
-    const { route } = useRouter();
-    const config = useConfig();
-    const { tags } = config.frontMatter;
-    const { resolvedTheme } = useTheme();
-
-    const similarArticles =
-      tags &&
-      allBlogs
-        .filter(
-          article =>
-            article.link !== route &&
-            (tags.length === 0 || article.tags?.some(tag => tags.includes(tag))),
-        )
-        .sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
-        .slice(0, 12)
-        // .sort(() => 0.5 - Math.random())
-        .slice(0, 4);
-
-    if (!route.startsWith('/blog/') || route.startsWith('/blog/tag')) {
-      return children as ReactElement;
-    }
-
-    return (
-      <>
-        <Article />
-        {children}
-        <Giscus
-          // ensure giscus is reloaded when client side route is changed
-          key={route}
-          repo="the-guild-org/website"
-          repoId="MDEwOlJlcG9zaXRvcnkxOTk3MTM1NzI="
-          category="Docs Discussions"
-          categoryId="DIC_kwDOC-djJM4CSZk-"
-          mapping="pathname"
-          theme={resolvedTheme}
-        />
-        <Newsletter className="mt-6 flex-col !gap-10" />
-        {!!similarArticles?.length && (
-          <>
-            <h3 className="text-center text-[28px] font-extrabold dark:text-[#FCFCFC]">
-              Similar articles
-            </h3>
-            <BlogCardList articles={similarArticles} className="xl:grid-cols-2" />
-          </>
-        )}
       </>
     );
   },
